Improve the precaching
The precaching of show information used to be broked and too agressive
in it's fetching. This new method saves the timestamp of when the show
info was last update and only updates when 30 days has passed. This
ensures that it's not done too often.

The reason for saveing the timestamp instead of using the HTTP caching
is that we need to be able to get fresh pages from the same urls that
give the show information so storing them in the cache for fast access
to show information is not an option.