Fix for Liveshows HTML update
The HTML had changed for the live shows page causing it to return zero
results. Fix this.

Also improve on the naming of the show which now inludes the time it is
broadcasted.

This fix also shows the next program that will become available as a
live show even though it is not yet being broadcast.